requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
alembic==1.17.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic-settings==2.11.0
PyJWT==2.10.1
python-jose==3.5.0
cryptography==46.0.3
passlib[bcrypt]==1.7.4
argon2-cffi==25.1.0
streamlit==1.50.0
email-validator==2.3.0
python-dotenv==1.1.1
requests==2.32.5